+++ database/migrations/2023_08_24_062608_create_wallets_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('wallets', function (Blueprint $table) {
             $table->id();
-            $table->double('balance')->default(0);
+            $table->unsignedBigInteger('balance')->default(0);
             $table->timestamp('activated_at')->nullable();
             $table->string('base58_check');
             $table->string('public_key');
